gpt5.6sol-xhigh

+++ src/crc.h
@@ -7,7 +7,8 @@
 extern const uint16_t crc16_table[256];
 
 static inline uint16_t crc16_update(uint16_t crc, uint8_t byte) {
-    return (crc << 8) ^ crc16_table[((crc >> 8) ^ byte) & 0xFF];
+    return (uint16_t)((crc << 8) ^
+                      crc16_table[((crc >> 8) ^ byte) & 0xFF]);
 }
 
 static inline uint16_t crc16(const uint8_t *data, size_t len, uint16_t init) {
